requirements.txt
streamlit
torch>=2.0
torchvision
pillow
transformers>=4.41.0
sentencepiece
accelerate
easyocr
opencv-python
numpy
nltk
spacy
pyttsx3
soundfile
langdetect
pypdf
pdf2image
pytz
regex
scikit-learn
rapidfuzz
markdown
